doxbox-master/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
crtsh==0.2.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-GoogleMaps==0.2.6
flask-nav==0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pygeoip==0.3.2
PyYAML==5.1
requests==2.21.0
SQLAlchemy==1.3.3
urllib3==1.24.3
visitor==0.1.3
Werkzeug==0.15.3
whois==0.7
WTForms==2.2.1
